Propose simple approach for migration discovery
This is not intended to be merged, just to start a discussion.

Although, if we do agree on this approach, it would be worth keeping it
as a design document or an ADR or something of the sort.

We have almost everything we need to greatly simplify the discovery of
migration scripts. We just need to fix version tagging in the Konflux
release pipeline [1] and in the build-definitions push script [2].

[1]: https://github.com/konflux-ci/release-service-catalog/tree/development/pipelines/managed/push-tekton-task-bundles-to-external-registry
[2]: https://github.com/konflux-ci/build-definitions/blob/main/hack/build-and-push.sh

Signed-off-by: Adam Cmiel <acmiel@redhat.com>