Handle invalid migration version
When a task has an invalid version that does not match versioning
expected by pipeline-migration-tool, it fails.

Catch the InvalidVersion exception in the Resolver.resolve() which is
ran when resolving migrations.
All the errors there are currently propagated as a MigrationResolveError
and if any occurs, it gets propagated above and stops ungracefully.
Catch the InvalidVersion exception and log a warning, but continue.

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>